chore: extract more types
