+++ coordinator/src/routes.rs
@@ -5,7 +5,6 @@ use crate::orderbook::routes::get_orders;
 use crate::orderbook::routes::post_order;
 use crate::orderbook::routes::put_order;
 use crate::orderbook::routes::websocket_handler;
-use crate::orderbook::routes::OrderbookMsg;
 use crate::AppError;
 use axum::extract::Path;
 use axum::extract::State;
@@ -19,6 +18,7 @@ use coordinator_commons::TradeParams;
 use diesel::r2d2::ConnectionManager;
 use diesel::r2d2::Pool;
 use diesel::PgConnection;
+use orderbook_commons::OrderbookMsg;
 use serde::Deserialize;
 use serde::Serialize;
 use std::collections::HashMap;
